protect stock routes

+++ server/controllers/stockController.js
@@ -1,13 +1,15 @@
 const Model = require('../models/Stock.Model');
+const User = require('../models/Users.Model');
 
 // @desc   Update stock
 // @route  POST /stock/update_stock
-// @access Public
+// @access Private
 const UpdateStock = async (req,res)=>{
     const data = new Model({
         item_name: req.body.item_name,
         item_quantity: req.body.item_quantity,
-        item_price: req.body.item_price
+        item_price: req.body.item_price,
+        user: req.user.id
     })
     try{
         const savedData = await data.save();
@@ -20,10 +22,10 @@ const UpdateStock = async (req,res)=>{
 
 // @desc   Get all stock
 // @route  GET /stock/get_stock
-// @access Public
+// @access Private
 const GetAllStock = async (req,res)=>{
     try{
-        const data = await Model.find();
+        const data = await Model.find({ user: req.user.id });
         res.status(200).json(data);
     }
     catch(err){
@@ -34,9 +36,22 @@ const GetAllStock = async (req,res)=>{
 
 // @desc   Get stock by id
 // @route  GET /stock/get_stock_by_id/:id
-// @access Public
+// @access Private
 const GetStockById = async (req,res)=>{
     try{
+        const id = req.params.id;
+        const user = await User.findById(req.user.id);
+
+        if(!user){
+            return res.status(404).json({message: "User not found"});
+        }
+
+        const itemOwner = await Model.findById(id);
+
+        if(user._id != itemOwner.user.toString()){
+            return res.status(401).json({message: "Unauthorized"});
+        }
+
         const data = await Model.findById(req.params.id);
         res.status(200).json(data);
     }
@@ -47,15 +62,27 @@ const GetStockById = async (req,res)=>{
 
 // @desc   Update stock by id
 // @route  PATCH /stock/update_stock_by_id/:id
-// @access Public
+// @access Private
 const UpdateStockById = async (req,res)=>{
     try{
         const id = req.params.id;
         const updated_data = req.body;
-        const options = {new: true};
 
-        const result = await Model.findByIdAndUpdate(id, updated_data, options);
+        const user = await User.findById(req.user.id);
+        const itemOwner = await Model.findById(id);
+
+        if(!user){
+            return res.status(404).json({message: "User not found"});
+        }
+
+        if(user._id != itemOwner.user.toString()){
+            return res.status(401).json({message: "Unauthorized"});
+        }
+        
+
+        const result = await Model.findByIdAndUpdate(id, updated_data);
         res.status(200).send(result);
+
     }
     catch(err){
         res.status(400).json({message: err});
@@ -64,10 +91,21 @@ const UpdateStockById = async (req,res)=>{
 
 // @desc   Delete stock by id
 // @route  DELETE /stock/delete_stock_by_id/:id
-// @access Public
+// @access Private
 const DeleteStockById = async (req,res)=>{
     try{
         const id = req.params.id;
+
+        const user = await User.findById(req.user.id);
+        const itemOwner = await Model.findById(id);
+
+        if(!user){
+            return res.status(404).json({message: "User not found"});
+        }
+
+        if(user._id != itemOwner.user.toString()){
+            return res.status(401).json({message: "Unauthorized"});
+        }
         const result = await Model.findByIdAndDelete(id);
         res.status(200).send(`Document with name ${result.item_name} deleted`);
     }
